+++ R/cypher_query.neo4j_engine.R
@@ -27,13 +27,16 @@ stitch_vectors <- function(x) {
 	}
 }
 
-########### Public functions ###########
-
-#' @export
-#' @importFrom neo2R cypher
-#' @importFrom tibble tibble
-cypher_query.neo4j_engine <- function(engine, query, parameters = NULL, ...) {	#
-	res <- neo2R::cypher(engine$graph_conn, query = query, parameters = parameters, result = "graph")
+#' Process neo2R cypher to tbl_kgx
+#'
+#' Given a result from neo2R::cypher returning KGX-formatted nodes and edges,
+#' parse the result to generate a tbl_kgx object, attaching the provided engine.
+#'
+#' @param res The result from neo2R::cypher with result = "graph"
+#' @param engine The engine to attach to the returned graph
+#' @importFrom memoise memoise
+#' @return A tbl_kgx
+neo2r_to_kgx <- function(res, engine) {
 	relationship_ids_contained <- as.integer(unlist(res$paths))
 
 	res <- stitch_vectors(res)
@@ -103,13 +106,13 @@ cypher_query.neo4j_engine <- function(engine, query, parameters = NULL, ...) {	#
 		# sapply!
 		# edges_df[[prop_name]] <- sapply(res$relationships, function(edge) {
 		edges_df[[prop_name]] <- sapply(res$relationships, function(edge) {
-#				edge$properties[[prop_name]]
-				prop_value <- edge$properties[[prop_name]]
-				if(is.null(prop_value)) {
-					return(NA)
-				} else {
-					return(prop_value)
-				}
+			#				edge$properties[[prop_name]]
+			prop_value <- edge$properties[[prop_name]]
+			if(is.null(prop_value)) {
+				return(NA)
+			} else {
+				return(prop_value)
+			}
 		})
 	}
 
@@ -120,4 +123,53 @@ cypher_query.neo4j_engine <- function(engine, query, parameters = NULL, ...) {	#
 	g <- tbl_kgx(nodes_df, edges_df, attach_engine = engine)
 	attr(g, "relationship_ids") <- relationship_ids_contained
 	return(g)
+}
+
+internal_cypher_query <- function(engine, query, parameters = NULL, ...) {	#
+	if(length(query) == 1) {
+		res <- neo2R::cypher(engine$graph_conn, query = query, parameters = parameters, result = "graph")
+		return(neo2r_to_kgx(res, engine = engine))
+	} else {
+		res <- neo2R::multicypher(engine$graph_conn, queries = query, parameters = parameters, result = "graph")
+		graphs <- lapply(res, neo2r_to_kgx, engine = engine)
+		g <- tbl_kgx(nodes = data.frame())
+		for(g2 in graphs) {
+			suppressMessages(g <- tidygraph::graph_join(g, g2), classes = "message") # suppress joining info
+		}
+		return(g)
+	}
+}
+
+#internal_cypher_query_memoised <- memoise::memoise(internal_cypher_query)
+
+########### Public functions ###########
+
+#' @export
+#' @importFrom neo2R cypher
+#' @importFrom neo2R multicypher
+#' @importFrom tibble tibble
+#' @importFrom tidygraph graph_join
+cypher_query.neo4j_engine <- function(engine, query, parameters = NULL, ...) {	#
+
+	if(!is.null(engine$cache)) {
+		# ok, this is a bit wonky
+		# the engine stores its cache
+		# we create a memoized internal function using that cache
+		# and then we call the function
+		# BUT, the engine itself needs to be sent to the function,
+		# and if its cache keeps changing it wont memoize properly
+		# so we create a copy of the engine without a cache and use that
+		engine_copy <- engine
+		engine_copy$cache <- NULL
+
+		internal <- memoise::memoise(internal_cypher_query, cache = engine$cache)
+		res <- internal(engine_copy, query, parameters, ...)
+
+		# before we return, we reset the cache of the engine attached to the graph.
+		res$engine$cache <- engine$cache
+		return(res)
+	} else {
+		internal_cypher_query(engine, query, parameters, ...)
+	}
+
 }